+++ airgun/views/common.py
@@ -1,4 +1,3 @@
-import time
 from widgetastic.widget import (
     Checkbox,
     ConditionalSwitchableView,
@@ -13,8 +12,6 @@ from widgetastic.widget import (
 from widgetastic_patternfly import BreadCrumb, Button, Tab, TabWithDropdown
 from widgetastic_patternfly4.navigation import Navigation
 from widgetastic_patternfly4.ouia import Dropdown
-from widgetastic_patternfly4.ouia import PatternflyTable
-from widgetastic_patternfly4.ouia import Button as PF4Button
 
 from airgun.utils import get_widget_by_name, normalize_dict_values
 from airgun.widgets import (
@@ -384,71 +381,6 @@ class AddRemoveSubscriptionsView(AddRemoveResourcesView):
         table = SatSubscriptionsTable(
             locator=".//table", column_widgets={0: Checkbox(locator=".//input[@type='checkbox']")}
         )
-
-
-class NewAddRemoveResourcesView(View):
-    searchbox = PF4Search()
-    type = Dropdown(
-        locator='.//div[contains(@class, "All repositories") or'
-        ' contains(@aria-haspopup="listbox")]'
-    )
-    Status = Dropdown(
-        locator='.//div[contains(@class, "All") or contains(@aria-haspopup="listbox")]'
-    )
-    add_repo = PF4Button('OUIA-Generated-Button-secondary-2')
-    # Need to add kebab menu
-    table = PatternflyTable(
-        component_id='OUIA-Generated-Table-4',
-        column_widgets={
-            0: Checkbox(locator='.//input[@type="checkbox"]'),
-            'Type': Text('.//a'),
-            'Name': Text('.//a'),
-            'Product': Text('.//a'),
-            'Sync State': Text('.//a'),
-            'Content': Text('.//a'),
-            'Status': Text('.//a'),
-        },
-    )
-
-    def search(self, value):
-        """Search for specific available resource and return the results"""
-        self.searchbox.search(value)
-        # Tried following ways to wait for table to be displayed, only sleep worked
-        # Might need a before/after fill
-        wait_for(
-            lambda: self.table.is_displayed is True,
-            timeout=60,
-            delay=1,
-        )
-        time.sleep(3)
-        self.table.wait_displayed()
-        return self.table.read()
-
-    def add(self, value):
-        """Associate specific resource"""
-        self.search(value)
-        next(self.table.rows())[0].widget.fill(True)
-        self.add_repo.click()
-
-    def fill(self, values):
-        """Associate resource(s)"""
-        if not isinstance(values, list):
-            values = list((values,))
-        for value in values:
-            self.add(value)
-
-    def remove(self, value):
-        """Unassign some resource(s).
-        :param str or list values: string containing resource name or a list of
-            such strings.
-        """
-        self.search(value)
-        next(self.table.rows())[0].widget.fill(True)
-        self.remove_button.click()
-
-    def read(self):
-        """Read all table values from both resource tables"""
-        return self.table.read()
 
 
 class TemplateEditor(View):
